Retry harvesting on 500 and 503 server errors
We currently use an automated retry with exponential backoff when
API servers respond with certain errors codes (429 and 502).

Some errors (e.g.
https://app.honeybadger.io/projects/131491/faults/121882510) were
caused by other 5XX codes, which might be resolved if we retried them.

This PR adds retries for 500 (generic server error) and 503 (service
unavailable) codes to try to address that.

Ref #389