Store records harvested via Kafka in the database (#446)
This converts the SDR DatasetRecordSet into a single, persistent
set that is not deleted when the TransformLoadJob runs. Instead,
the SDR harvester updates records in that set as they are published,
in addition to indexing them immediately.

This prevents SDR records from getting deleted during Transform runs
since they are always "the most recent set".