Add additional logging to SdrConsumer
This adds more logging at the DEBUG level so that it's possible
to check every Kafka message coming in from purl-fetcher.

Also changes the logger to use the same one that racecar uses
internally, which provides timestamps on the log messages so that
it's easier to line things up when testing.