Refactor SdrConsumer to remove stored interstitial data
The old version involved a lot of temporary storage of data per
each message as instance variables that were overwritten, which
made it hard to keep track of values.

This parameterizes most of the methods so that the data flow is
more clear.

Also sets the DOI field when saving records to the database.