Fixes for SDR harvest deletion & skips (#492)
* Find the SDR DatasetRecordSet by provider only

* Don't use the job ID as load ID for SdrExtractJob

* Fix an issue with using SdrConsumer#update_item via console

* Refactor SdrConsumer to remove stored interstitial data

The old version involved a lot of temporary storage of data per
each message as instance variables that were overwritten, which
made it hard to keep track of values.

This parameterizes most of the methods so that the data flow is
more clear.

Also sets the DOI field when saving records to the database.

* Add SDR items from SdrExtractJob using the bare druid

This matches how the SdrConsumer adds items to the database;
neither should use the 'druid:' prefix for consistency.